Moved files, delete redundant files, fix for tests after changes
